Define stride_t as more portable off_t rather than long long.
git-svn-id: http://svn.openimageio.org/oiio/trunk@117 31904243-a364-42ac-a3db-9305396518ba

+++ src/include/imageio.h
@@ -75,7 +75,7 @@ const int IMAGEIO_VERSION = 10;
 
 /// Type we use for stride lengths.  It should be 64 bit on all 
 /// supported systems.
-typedef long long stride_t;
+typedef off_t stride_t;
